BinanceCoveBot/requirements.txt
telethon==1.34.0
binance-futures-connector==4.0.0
pyyaml
tinydb==4.8.0
python-dotenv

